requirements-langfuse.txt
langfuse>=3.0.0
opentelemetry-instrumentation-anthropic
